Add session data object and node_editor callbacks to r/w to it.
Preivously the session data was R/Wed directly to a disk file by the node editor backend.  This implements writing it to a string buffer in the session_data.  In this case, I am also R/Wing it to disk, but with this hooked up, I can in the future write a save function that is only called on user command (or auto-backup).