Serialize and De-serialize registry nodes
This also consolidates the individual node properties into one file, but for now the node location file from NodeEditor is in a second file.

Deserializing from save files adds node IDs in a new way, and as a side effect we had to mess around with the "next ID" variable, so that deserializing 5 nodes, then creating one node didn't blast away the first deserialized node.